moving files from bayesian directory
