+++ selection/reduced_optimization/barrier.py
