+++ selection/reduced_optimization/par_carved_reduced.py
@@ -2,9 +2,9 @@ import numpy as np
 import sys
 
 import regreg.api as rr
-from selection.reduced_optimization.lasso_reduced import nonnegative_softmax_scaled, neg_log_cube_probability
-from selection.bayesian.credible_intervals import projected_langevin
-from selection.reduced_optimization.par_random_lasso_reduced import log_likelihood
+from .lasso_reduced import nonnegative_softmax_scaled, neg_log_cube_probability
+from .credible_intervals import projected_langevin
+from .par_random_lasso_reduced import log_likelihood
 
 class smooth_cube_barrier(rr.smooth_atom):
