+++ selection/reduced_optimization/tests/crime_data_set.py
@@ -2,13 +2,13 @@ import os, numpy as np, pandas, statsmodels.api as sm
 import time
 import matplotlib.pyplot as plt
 import regreg.api as rr
-from selection.bayesian.initial_soln import selection
+from selection.reduced_optimization.initial_soln import selection
 from selection.randomized.api import randomization
 from selection.reduced_optimization.lasso_reduced import nonnegative_softmax_scaled, neg_log_cube_probability, selection_probability_lasso, \
     sel_prob_gradient_map_lasso, selective_inf_lasso
 
 
-NRTI =  pandas.read_table("/Users/snigdhapanigrahi/Results_bayesian/trimmed_communities.txt", sep='\s+')
+NRTI =  pandas.read_table("/Users/snigdhapanigrahi/Results_reduced_optimization/trimmed_communities.txt", sep='\s+')
 print("shape of data", NRTI.shape)
 
 NRTI = NRTI.as_matrix()
@@ -141,4 +141,4 @@ ax.legend((rects1[0], rects2[0]), ('Unadjusted', 'Adjusted'), loc='upper left')
 
 print('here')
 
-plt.savefig('/Users/snigdhapanigrahi/Results_bayesian/credible_crime.pdf', bbox_inches='tight')
+plt.savefig('/Users/snigdhapanigrahi/Results_reduced_optimization/credible_crime.pdf', bbox_inches='tight')
