+++ selection/reduced_optimization/tests/random_reduced_lasso_bayesian_model.py
@@ -4,7 +4,7 @@ import os
 import numpy as np
 import time
 import regreg.api as rr
-from selection.bayesian.initial_soln import selection
+from selection.reduced_optimization.initial_soln import selection
 from selection.tests.instance import logistic_instance, gaussian_instance
 
 #from selection.reduced_optimization.random_lasso_reduced import selection_probability_random_lasso, sel_inf_random_lasso
