+++ selection/reduced_optimization/tests/read_file.py
@@ -2,10 +2,10 @@ import glob
 import os, numpy as np, pandas, statsmodels.api as sm
 
 #path =r'/Users/snigdhapanigrahi/Results_freq_EQTL/sparsity_5/dim_1/dim_1'
-#path =r'/Users/snigdhapanigrahi/Results_bayesian/fixed_lasso/fixed_lasso'
+#path =r'/Users/snigdhapanigrahi/Results_reduced_optimization/fixed_lasso/fixed_lasso'
 
-path =r'/Users/snigdhapanigrahi/Results_bayesian/experiment_dual_0'
-#path =r'/Users/snigdhapanigrahi/Results_bayesian/bayesian_dual'
+path =r'/Users/snigdhapanigrahi/Results_reduced_optimization/experiment_dual_0'
+#path =r'/Users/snigdhapanigrahi/Results_reduced_optimization/bayesian_dual'
 allFiles = glob.glob(path + "/*.txt")
 
 list_ = []
