+++ selection/reduced_optimization/tests/test_selection_random_lasso.py
@@ -2,7 +2,7 @@ from __future__ import print_function
 import numpy as np
 import time
 import regreg.api as rr
-from selection.bayesian.initial_soln import selection
+from selection.reduced_optimization.initial_soln import selection
 from selection.tests.instance import logistic_instance, gaussian_instance
 
 #from selection.reduced_optimization.random_lasso_reduced import log_likelihood, selection_probability_random_lasso, sel_inf_random_lasso
